requirements.txt
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.3
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dnspython==2.2.1
fastapi==0.73.0
flower==1.0.0
h11==0.12.0
httpcore==0.14.7
httpx==0.21.1
humanize==4.3.0
idna==3.3
kombu==5.2.4
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.31
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.17.1
tornado==6.2
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
wrapt==1.14.1
xmltodict==0.13.0
